requirements.txt
requests
python-dotenv
attrs==18.2.0
pdflatex
jinja2
